combine common faheader and header elements into shared file in themes
