+++ themes/ihlc/faheader.inc.php
@@ -57,214 +57,8 @@ else
    $_ARCHON->PublicInterface->Title = $_ARCHON->PublicInterface->Title ? $_ARCHON->PublicInterface->Title . ' - ' . 'Archon' : 'Archon';
 }
 
-//adds to breadcrumbs
-$_ARCHON->PublicInterface->addNavigation('Manuscript Collections Database', 'index.php', true);
-$_ARCHON->PublicInterface->addNavigation('IHLC', 'https://www.library.illinois.edu/ihx/', true);
-
-header('Content-type: text/html; charset=UTF-8');
+require('commonheader.inc.php');         
 ?>
-<!DOCTYPE html>
-<html lang="en">
-   <head>
-      <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-      <meta name="og:site_name" content="Illinois History and Lincoln Collections Manuscript Collections Database"/>
-	   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-      <title><?php echo(strip_tags($_ARCHON->PublicInterface->Title)); ?></title>
-      <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->Theme); ?>/style.css" />
-      <link rel="stylesheet" type="text/css" href="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.css" />
-      <link rel="icon" type="image/ico" href="<?php echo($_ARCHON->PublicInterface->ImagePath); ?>/favicon.ico"/>
-      <!--[if IE]>
-        <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->Theme); ?>/ie.css" />
-        <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.ie.css" />
-      <![endif]-->
-<?php echo($_ARCHON->getJavascriptTags('jquery.min')); ?>
-<?php echo($_ARCHON->getJavascriptTags('jquery-ui.custom.min')); ?>
-<?php echo($_ARCHON->getJavascriptTags('jquery-expander')); ?>
-      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jquery.hoverIntent.js"></script>
-      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.js"></script>
-      
-      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jquery.scrollTo-min.js"></script>
-<?php echo($_ARCHON->getJavascriptTags('archon')); ?>
-      <script type="text/javascript">
-         /* <![CDATA[ */
-         imagePath = '<?php echo($_ARCHON->PublicInterface->ImagePath); ?>';
-         $(document).ready(function() {            
-            $('div.listitem:nth-child(even)').addClass('evenlistitem');
-            $('div.listitem:last-child').addClass('lastlistitem');
-            $('#locationtable tr:nth-child(odd)').addClass('oddtablerow');
-            $('.expandable').expander({
-               slicePoint:       1000,              // make expandable if over this x chars
-               widow:            100,              // do not make expandable unless total length > slicePoint + widow
-               expandPrefix:     '',         // text to come before the expand link
-               expandText:         'READ MORE',  //text to use for expand link
-               expandEffect:     'fadeIn',         // or slideDown
-               expandSpeed:      700,              // in milliseconds
-               collapseTimer:    0,                // milliseconds before auto collapse; default is 0 (don't re-collapse)
-               userCollapseText: 'COLLAPSE TEXT'      // text for collapse link
-            });
-         });
-
-         function js_highlighttoplink(selectedSpan)
-         {
-            $('.currentBrowseLink').toggleClass('browseLink').toggleClass('currentBrowseLink');
-            $(selectedSpan).toggleClass('currentBrowseLink');
-            $(selectedSpan).effect('highlight', {}, 400);
-         }
-
-         $(window).load(function() {<?php echo($_ARCHON->PublicInterface->Header->OnLoad); ?>});
-         $(window).unload(function() {<?php echo($_ARCHON->PublicInterface->Header->OnUnload); ?>});
-         /* ]]> */
-      </script>
-<?php
-      if($_ARCHON->PublicInterface->Header->Message && $_ARCHON->PublicInterface->Header->Message != $_ARCHON->Error)
-      {
-         $message = $_ARCHON->PublicInterface->Header->Message;
-      } ?>
-   </head>
-   <body>
-<?php
-      $_ARCHON->PublicInterface->outputGoogleAnalyticsCode();
-
-
-      if($message)
-      {
-         echo("<div class='message'>" . encode($message, ENCODE_HTML) . "</div>\n");
-      }
-?>
-      <div id='top'>
-		<div id="headerbar">
-			<div id="topnavblock">
-      
-      <a href='http://illinois.edu'>University of Illinois at Urbana-Champaign</a> > <a href='http://library.illinois.edu'>Library</a> > <a href='http://www.library.illinois.edu/ihx'>IHLC</a> > <a href="http://www.library.illinois.edu/ihx/archon/">Manuscript Collections Database</a>
-      </div>
-			<div id="researchblock">
-<?php
-                  if($_ARCHON->Security->isAuthenticated())
-                  {
-                     echo("<span class='bold'>Welcome, " . $_ARCHON->Security->Session->User->toString() . "</span><br/>");
-
-                     $logoutURI = preg_replace('/(&|\\?)f=([\\w])*/', '', $_SERVER['REQUEST_URI']);
-                     $Logout = (encoding_strpos($logoutURI, '?') !== false) ? '&amp;f=logout' : '?f=logout';
-                     $strLogout = encode($logoutURI, ENCODE_HTML) . $Logout;
-                     echo("<a href='$strLogout'>Logout</a>");
-                  }
-                  else
-                  {
-                     echo("<a href='#' onclick='$(window).scrollTo(\"#archoninfo\"); if($(\"#userlogin\").is(\":visible\")) $(\"#loginlink\").html(\"Log In (Registered Researchers or Staff)\"); else $(\"#loginlink\").html(\"Hide\"); $(\"#userlogin\").slideToggle(\"normal\"); $(\"#ArchonLoginField\").focus(); return false;'>Log In</a>");
-                  }
-
-                  if(!$_ARCHON->Security->userHasAdministrativeAccess())
-                  {
-                     $emailpage = defined('PACKAGE_COLLECTIONS') ? "collections/research" : "core/contact";
-
-                     /*echo(" | <a href='?p={$emailpage}&amp;f=email&amp;referer=" . urlencode($_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI']) . "'>Contact Us</a>");*/
-					 
-					 echo(" | <a href='mailto:ihlc@library.illinois.edu'>Contact Us</a>");
-
-                     if($_ARCHON->Security->isAuthenticated())
-                     {
-                        echo(" | <a href='?p=core/account&amp;f=account'>My Account</a>");
-                     }
-                     /*if(defined('PACKAGE_COLLECTIONS'))
-                     {
-                        $_ARCHON->Security->Session->ResearchCart->getCart();
-                        $EntryCount = $_ARCHON->Security->Session->ResearchCart->getCartCount();
-
-                        echo(" | <a href='?p=collections/research&amp;f=cart&amp;referer=" . urlencode($_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI']) . "'>View Cart (<span id='cartcount'>$EntryCount</span>)</a>");
-                     }*/
-                  }
-?>
-               </div>
-		</div>
-		
-		<div id="headercenter">
-			<div id="headertitleblock">
-				<a href='http://www.library.illinois.edu/ihx/index.html'>Illinois History and Lincoln Collections</a> | <a href='./index.php'>Manuscript Collections</a>
-			</div>
-			<div id="searchwrapper">
-            <!--<div id="logo"><a href="index.php" ><img src="<?php echo($_ARCHON->PublicInterface->ImagePath); ?>/logo.gif" alt="logo" /></a><br/>Archives</div> -->
-            <div id="searchblock">
-               <form action="index.php" accept-charset="UTF-8" method="get" onsubmit="if(!this.q.value) { alert('Please enter search terms.'); return false; } else { return true; }">
-                  <div>
-                     <input type="hidden" name="p" value="core/search" />
-                     <label for='q'></label>
-					 <input type="text" size="35" maxlength="150" name="q" id="q" value="<?php echo(encode($_ARCHON->QueryString, ENCODE_HTML)); ?>" tabindex="100" placeholder="Search collection summaries"/>
-                     <input type="submit" value="Find" tabindex="300" class='button' title="Search" /> 
-					 
-<?php
-      if(defined('PACKAGE_COLLECTIONS') && CONFIG_COLLECTIONS_SEARCH_BOX_LISTS)
-      {
-?>
-                     <input type="hidden" name="content" value="1" />
-                     <?php
-                  }
-                     ?>
-                  </div></form></div>
-				  
-				<div id="pdfsearchlink"><a class='bold' style='color:white' href='?p=core/index'>Return to main search page</a></div>
-         </div>
-		</div>
-		
-		<div id="headernav">
-			<?php
-                  $arrP = explode('/', $_REQUEST['p']);
-                  $TitleClass = $arrP[0] == 'collections' && $arrP[1] != 'classifications' ? 'currentBrowseLink' : 'browseLink';
-                  $ClassificationsClass = $arrP[1] == 'classifications' ? 'currentBrowseLink' : 'browseLink';
-                  $SubjectsClass = $arrP[0] == 'subjects' ? 'currentBrowseLink' : 'browseLink';
-                  $CreatorsClass = $arrP[0] == 'creators' ? 'currentBrowseLink' : 'browseLink';
-                  $DigitalLibraryClass = $arrP[0] == 'digitallibrary' ? 'currentBrowseLink' : 'browseLink';
-            ?>
-                  <div id="browsebyblock">
-            
-					 <span id="browsebyspan">
-                        Browse by:
-                     </span>
-					 
-                     <span class="<?php echo($TitleClass); ?>">
-                        <a href="?p=collections/collections" onclick="js_highlighttoplink(this.parentNode); return true;">Collection Titles</a>
-                     </span>
-
-                     <span class="<?php echo($SubjectsClass); ?>">
-                        <a href="?p=subjects/subjects" onclick="js_highlighttoplink(this.parentNode); return true;">Subjects</a>
-                     </span>
-					 
-                     <span class="<?php echo($CreatorsClass); ?>">
-                        <a href="?p=creators/creators" onclick="js_highlighttoplink(this.parentNode); return true;">Persons or Organizations</a>
-                     </span>
-					 
-                     <!--<span class="<?php echo($DigitalLibraryClass); ?>">
-                        <a href="?p=digitallibrary/digitallibrary" onclick="js_highlighttoplink(this.parentNode); return true;">Images</a>
-                     </span> -->
-
-                  </div>
-		</div>
- 
-    </div>
-
-
-               <div id="breadcrumbblock">
-<?php echo($_ARCHON->PublicInterface->createNavigation()); ?>
-               </div>
-               <div id="breadcrumbclearblock">.</div>
-
-               <script type="text/javascript">
-                  /* <![CDATA[ */
-                  if ($.browser.msie && parseInt($.browser.version, 10) <= 8){
-                     $.getScript('packages/core/js/jquery.corner.js', function(){
-                        $("#searchblock").corner("5px");
-                        $("#browsebyblock").corner("tl 10px");
-
-                        $(function(){
-                           $(".bground").corner("20px");
-                           $(".mdround").corner("10px");
-                           $(".smround").corner("5px");
-                           $("#dlsearchblock").corner("bottom 10px");
-                        });
-                     });
-                  }
-                  /* ]]> */
-               </script>         
-
                <!-- Begin Navigation Bar for Finding Aid View -->
 
                <div id='left'>
@@ -300,38 +94,45 @@ header('Content-type: text/html; charset=UTF-8');
                   {
  ?> <p><a href="#admininfo" tabindex="700">Administrative Information</a></p> <?php } ?>
             <?php if(!empty($objCollection->Content))
-                  {
- ?> <p><a href="#boxfolder" tabindex="800">Detailed Description</a></p><?php
-                  }
+            {
+?> <p><a href="#boxfolder" tabindex="800">Detailed Description</a></p><?php
+            
 
-                  foreach($objCollection->Content as $ID => $objContent)
-                  {
-                     if(!$objContent->ParentID && trim($objContent->Title))
-                     {
-                        echo("<p class='faitemcontent'><a href='?p=collections/findingaid&amp;id=$objCollection->ID&amp;q=$_ARCHON->QueryStringURL&amp;rootcontentid=$ID#id$ID'>" . $objContent->getString('Title') . "</a></p>\n");
-                     }
-                     else if(!$objContent->ParentID)
-                     {
-                        $LevelContainerString = $objContent->LevelContainer ? $objContent->LevelContainer->getString('LevelContainer') : '';
-                        echo("<p class='faitemcontent'><a href='?p=collections/findingaid&amp;id=$objCollection->ID&amp;q=$_ARCHON->QueryStringURL&amp;rootcontentid=$ID#id$ID'>{$LevelContainerString} " . formatNumber($objContent->LevelContainerIdentifier) . "</a></p>\n");
-                     }
-                  }
-            ?>
-                  <form action="index.php" accept-charset="UTF-8" method="get" onsubmit="if(!this.q.value) { alert('Please enter search terms.'); return false; } else { return true; }">
-                     <div id="fasearchblock">
-                        <input type="hidden" name="p" value="core/search" />
-                        <input type="hidden" name="flags" value="<?php echo(SEARCH_COLLECTIONCONTENT); ?>" />
-                        <input type="hidden" name="collectionid" value="<?php echo($objCollection->ID); ?>" />
-                        <input type="hidden" name="content" value="1" />
-                        <label for='q'></label><input type="text" size="20" maxlength="150" name="q" id="q" value="<?php echo(encode($_ARCHON->QueryString, ENCODE_HTML)); ?>" tabindex="100" /><br/>
-                        <input type="submit" value="Search This Finding Aid" tabindex="200" class='button' title="Search Box List" />
-                     </div>
-                  </form>
+            foreach($objCollection->Content as $ID => $objContent)
+            {
+               if(!$objContent->ParentID && trim($objContent->Title))
+               {
+                  echo("<p class='faitemcontent'><a href='?p=collections/findingaid&amp;id=$objCollection->ID&amp;q=$_ARCHON->QueryStringURL&amp;rootcontentid=$ID#id$ID'>" . $objContent->getString('Title') . "</a></p>\n");
+               }
+               else if(!$objContent->ParentID)
+               {
+                  $LevelContainerString = $objContent->LevelContainer ? $objContent->LevelContainer->getString('LevelContainer') : '';
+                  echo("<p class='faitemcontent'><a href='?p=collections/findingaid&amp;id=$objCollection->ID&amp;q=$_ARCHON->QueryStringURL&amp;rootcontentid=$ID#id$ID'>{$LevelContainerString} " . formatNumber($objContent->LevelContainerIdentifier) . "</a></p>\n");
+               }
+            }
+      ?>
+            <form action="index.php" accept-charset="UTF-8" method="get" onsubmit="if(!this.q.value) { alert('Please enter search terms.'); return false; } else { return true; }">
+               <div id="fasearchblock">
+                  <input type="hidden" name="p" value="core/search" />
+                  <input type="hidden" name="flags" value="<?php echo(SEARCH_COLLECTIONCONTENT); ?>" />
+                  <input type="hidden" name="collectionid" value="<?php echo($objCollection->ID); ?>" />
+                  <input type="hidden" name="content" value="1" />
+                  <input type="text" size="20" title="search" maxlength="150" name="q" id="q2" value="<?php echo(encode($_ARCHON->QueryString, ENCODE_HTML)); ?>" tabindex="100" /><br/>
+                  <input type="submit" value="Search Box List" tabindex="200" class='button' title="Search Box List" />
+               </div>
+            </form>
 <?php
-                  if(defined('PACKAGE_RESEARCH'))
-                  {
-                     echo("<hr/><p class='center' style='font-weight:bold'><a href='?p=research/research&amp;f=email&amp;referer=" . urlencode($_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI']) . "'>Contact us about this collection</a></p>");
-                  }
+            }//end if statement for if collection content is not empty
+
+            if($objCollection->OtherURL){
+
+               if(strtolower(substr($objCollection->getString('OtherURL'),-3))=="pdf"){
+                  echo( '<p><a href="#pdf-fa" tabindex="800">PDF Box/Folder List</a></p>');
+               } elseif($objCollection->getString('OtherURL')=="http://hdl.handle.net/2027/uc1.b4265910" OR $objCollection->getString('OtherURL')=="https://archive.org/details/guidetoheinricha00univ" OR $objCollection->getString('OtherURL')=="https://hdl.handle.net/2027/uiuo.ark:/13960/t9086670x"){
+				   echo( '<p><a href="#pdf-fa" tabindex="800">Embedded Published Finding Aid</a></p>');
+			   }
+
+            }
 ?>
 
          </div>
