+++ themes/library_web/faheader.inc.php
@@ -56,158 +56,8 @@ else
    $_ARCHON->PublicInterface->Title = $_ARCHON->PublicInterface->Title ? $_ARCHON->PublicInterface->Title . ' - ' . 'Archon' : 'Archon';
 }
 
-$_ARCHON->PublicInterface->addNavigation('Holdings', 'index.php', true);
-$_ARCHON->PublicInterface->addNavigation('University Archives', 'https://archives.library.illinois.edu', true);
-
-header('Content-type: text/html; charset=UTF-8');
-?>
-
-<!DOCTYPE html>
-<html lang="en">
-   <head>
-      <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-      <meta name="og:site_name" content="University of Illinois Archives Holdings Database"/>
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-	  <title><?php echo(strip_tags($_ARCHON->PublicInterface->Title)); ?></title>
-      <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->Theme); ?>/style.css" />
-      <link rel="stylesheet" type="text/css" href="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.css" />
-      <link rel="stylesheet" type="text/css" href="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jgrowl/jquery.jgrowl.css" />
-      <link rel="icon" type="image/ico" href="<?php echo($_ARCHON->PublicInterface->ImagePath); ?>/favicon.ico"/>
-      <!--[if lte IE 7]>
-        <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->Theme); ?>/ie.css" />
-        <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.ie.css" />
-      <![endif]-->
-      <?php echo($_ARCHON->getJavascriptTags('jquery.min')); ?>
-      <?php echo($_ARCHON->getJavascriptTags('jquery-ui.custom.min')); ?>
-      <?php echo($_ARCHON->getJavascriptTags('jquery-expander')); ?>
-      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jquery.hoverIntent.js"></script>
-      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.js"></script>
-      
-      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jquery.scrollTo-min.js"></script>
-      <?php echo($_ARCHON->getJavascriptTags('jquery.jgrowl.min')); ?>
-      <?php echo($_ARCHON->getJavascriptTags('archon')); ?>
-      <script type="text/javascript">
-         /* <![CDATA[ */
-         imagePath = '<?php echo($_ARCHON->PublicInterface->ImagePath); ?>';
-         $(document).ready(function() {          
-            $('div.listitem:nth-child(even)').addClass('evenlistitem');
-            $('div.listitem:last-child').addClass('lastlistitem');
-            $('#locationtable tr:nth-child(odd)').addClass('oddtablerow');
-            $('.expandable').expander({
-               slicePoint:       600,              // make expandable if over this x chars
-               widow:            100,              // do not make expandable unless total length > slicePoint + widow
-               expandPrefix:     '. . . ',         // text to come before the expand link
-               expandText:         '[read more]',  //text to use for expand link
-               expandEffect:     'fadeIn',         // or slideDown
-               expandSpeed:      700,              // in milliseconds
-               collapseTimer:    0,                // milliseconds before auto collapse; default is 0 (don't re-collape)
-               userCollapseText: '[collapse]'      // text for collaspe link
-            });
-         });
-
-         function js_highlighttoplink(selectedSpan)
-         {
-            $('.currentBrowseLink').toggleClass('browseLink').toggleClass('currentBrowseLink');
-            $(selectedSpan).toggleClass('currentBrowseLink');
-            $(selectedSpan).effect('highlight', {}, 300);
-         }
-
-         $(window).load(function() {<?php echo($_ARCHON->PublicInterface->Header->OnLoad); ?>});
-         $(window).unload(function() {<?php echo($_ARCHON->PublicInterface->Header->OnUnload); ?>});
-         /* ]]> */
-      </script>
-      <?php
-      if($_ARCHON->PublicInterface->Header->Message && $_ARCHON->PublicInterface->Header->Message != $_ARCHON->Error)
-      {
-         $message = $_ARCHON->PublicInterface->Header->Message;
-      } ?>
-      <script src="https://use.fontawesome.com/77e525f0e0.js"></script>
-   </head>
-   <body>
-      <?php
-      $_ARCHON->PublicInterface->outputGoogleAnalyticsCode();
-
-
-      if($message)
-      {
-         echo("<div class='message'>" . encode($message, ENCODE_HTML) . "</div>\n");
-      }
-      ?>
-      <div id='top'>
-         <div id="logosearchwrapper">
-         <div id="sitetitleblock"><a href="https://archives.library.illinois.edu">University of Illnois Archives</a> &ndash;> <a href="index.php">Holdings Database</a></div>
-            <div id="searchblock">
-               <form action="index.php" accept-charset="UTF-8" method="get" onsubmit="if(!this.q.value) { alert('Please enter search terms.'); return false; } else { return true; }">
-                  <div>
-                     <input type="hidden" name="p" value="core/search" />
-                     <input type="text" size="25" title="search" maxlength="150" name="q" id="qfa" value="<?php echo(encode($_ARCHON->QueryString, ENCODE_HTML)); ?>" tabindex="100" />
-                     <input type="submit" value="Search" tabindex="300" class='button' title="Search" />
-                     <?php
-                     if(defined('PACKAGE_COLLECTIONS') && CONFIG_COLLECTIONS_SEARCH_BOX_LISTS)
-                     {
-                     ?>
-                        <input type="hidden" name="content" value="0"/>
-<?php
-                     }
-?>
-                  </div>
-               </form>
-            </div>
-         </div>
-
-<?php
-                     $arrP = explode('/', $_REQUEST['p']);
-                     $TitleClass = $arrP[0] == 'collections' && $arrP[1] != 'classifications' ? 'currentBrowseLink' : 'browseLink';
-                     $ClassificationsClass = $arrP[1] == 'classifications' ? 'currentBrowseLink' : 'browseLink';
-                     $SubjectsClass = $arrP[0] == 'subjects' ? 'currentBrowseLink' : 'browseLink';
-                     $CreatorsClass = $arrP[0] == 'creators' ? 'currentBrowseLink' : 'browseLink';
-                     $DigitalLibraryClass = $arrP[0] == 'digitallibrary' ? 'currentBrowseLink' : 'browseLink';
-?>
-    <div id="browsebyblock">
-            <div id="browseblockcontent">
-            <span class="<?php echo($ClassificationsClass); ?>">
-               <a href="?p=collections/classifications" onclick="js_highlighttoplink(this.parentNode); return true;">Campus Units/Record Groups</a>
-            </span>
-
-            <span class="<?php echo($TitleClass); ?>">
-               <a href="?p=collections/collections" onclick="js_highlighttoplink(this.parentNode); return true;">Physical Collections: A-Z</a>
-            </span>
-			 <span class="<?php echo($DigitalLibraryClass); ?>">
-               <a href="?p=digitallibrary/digitallibrary" onclick="js_highlighttoplink(this.parentNode); return true;">Digital Materials: A-Z</a>
-            </span>
-           
-            <span class="<?php echo($SubjectsClass); ?>">
-               <a href="?p=subjects/subjects" onclick="js_highlighttoplink(this.parentNode); return true;">Subject Headings</a>
-            </span>
-            <span class="<?php echo($CreatorsClass); ?>">
-               <a href="?p=creators/creators" onclick="js_highlighttoplink(this.parentNode); return true;">Record Creators</a>
-            </span>
-           </div>
-         </div>
-      </div>
-
-                  <div id="breadcrumbblock">
-<?php echo($_ARCHON->PublicInterface->createNavigation()); ?>
-                  </div>
-                  <div id="breadcrumbclearblock">.</div>
-
-                  <script type="text/javascript">
-                     /* <![CDATA[ */
-                     if ($.browser.msie && parseInt($.browser.version, 10) <= 8){
-                        $.getScript('packages/core/js/jquery.corner.js', function(){
-                           $("#searchblock").corner("5px");
-                           $("#browsebyblock").corner("tl 10px");
-
-                           $(function(){
-                              $(".bground").corner("20px");
-                              $(".mdround").corner("10px");
-                              $(".smround").corner("5px");
-                              $("#dlsearchblock").corner("bottom 10px");
-                           });
-                        });
-                     }
-                     /* ]]> */
-                  </script>            
+require('commonheader.inc.php');
+?>           
 
                   <!-- Begin Navigation Bar for Finding Aid View -->
 
@@ -296,7 +146,48 @@ header('Content-type: text/html; charset=UTF-8');
             <?php
                      if(defined('PACKAGE_COLLECTIONS'))
                      {
-                        echo("<hr/><p class='center' style='font-weight:bold'><a href='?p=collections/research&amp;f=email&amp;referer=" . urlencode($_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI']) . "'>Contact us about this collection</a></p>");
+                        echo("<hr/><p class='center' style='font-weight:bold'>");
+                        if ($objCollection->RepositoryID == 2)
+                        {
+                           $emailmailto = "sousa@illinois.edu";
+                           $emailsubject="Reference inquiry for Sousa Archives";
+                           $emailbody="%0D---Please type your message above this line---%0DThis email was sent from: https://".urlencode($_SERVER['HTTP_HOST']). urlencode($_SERVER['REQUEST_URI']);
+                           
+                        }
+                        else 
+                        {
+                           $emailmailto = "illiarch@illinois.edu";
+                           $emailsubject="Inquiry: University Archives Finding Aid";
+                           $emailReferralPage = "https://".urlencode($_SERVER['HTTP_HOST']). urlencode($_SERVER['REQUEST_URI']);
+                           $emailbody="%0D---Please type your message above this line---%0DReferral page: ".$emailReferralPage;
+                        }
+                        if($objCollection->Classification) {
+                           $emailsubject .= " (RS ".$objCollection->Classification->toString(LINK_NONE, true, false, true, false);
+                           $emailsubject .= "/".$objCollection->getString('CollectionIdentifier').")";
+                        }
+                        if($objCollection->RepositoryID == 2){
+                           echo("<a href='mailto:".$emailmailto."?subject=". $emailsubject .'&body='.$emailbody."'>");
+                        } else {
+                           echo("<a href='https://archives.library.illinois.edu/email-ahx.php?this_page=".$emailReferralPage."' target='_blank'>");
+                        }
+                        echo("Email us about these ");
+                              
+                        if($objCollection->MaterialType == 'Official Records--Non-University' || $objCollection->MaterialType == 'Official Records')
+                        {
+                           echo ('records');
+                        }
+                           elseif($objCollection->MaterialType == 'Publications')
+                           {
+                              echo ('publications');
+                           }
+                           else
+                           {
+                              echo ('papers');
+                           }
+               
+                           echo("</a>");        
+                        
+                        echo("</p>");
                      }
             ?>
