+++ themes/sousa2/commonheader.inc.php
@@ -1,0 +1,210 @@
+<?php
+/**
+ * Common elements of the regular header and the finding aid header.
+ */
+
+
+$_ARCHON->PublicInterface->addNavigation('Archon', 'index.php', true);
+
+//header('Content-type: text/html; charset=UTF-8');
+?>
+
+<!DOCTYPE html>
+<html lang="en">
+   <head>
+      <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+      <meta name="viewport" content="width=device-width, initial-scale=1.0">
+      <title><?php echo(strip_tags($_ARCHON->PublicInterface->Title)); ?></title>
+      <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->Theme); ?>/style.css?v=20230315" />
+      <link rel="stylesheet" type="text/css" href="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.css" />
+      <link rel="stylesheet" type="text/css" href="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jgrowl/jquery.jgrowl.css" />
+
+      <link rel="icon" type="image/ico" href="<?php echo($_ARCHON->PublicInterface->ImagePath); ?>/archon.ico"/>
+      <!--[if lte IE 7]>
+        <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->Theme); ?>/ie.css" />
+        <link rel="stylesheet" type="text/css" href="themes/<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.ie.css" />
+      <![endif]-->
+      <?php echo($_ARCHON->getJavascriptTags('jquery.min')); ?>
+      <?php echo($_ARCHON->getJavascriptTags('jquery-ui.custom.min')); ?>
+      <?php echo($_ARCHON->getJavascriptTags('jquery-expander')); ?>
+      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jquery.hoverIntent.js"></script>
+      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/cluetip/jquery.cluetip.js"></script>
+      <script type="text/javascript" src="<?php echo($_ARCHON->PublicInterface->ThemeJavascriptPath); ?>/jquery.scrollTo-min.js"></script>
+      <?php echo($_ARCHON->getJavascriptTags('jquery.jgrowl.min')); ?>
+      <?php echo($_ARCHON->getJavascriptTags('archon')); ?>
+
+      <script type="text/javascript">
+         /* <![CDATA[ */
+         imagePath = '<?php echo($_ARCHON->PublicInterface->ImagePath); ?>';
+         $(document).ready(function() {
+            $('div.listitem:nth-child(even)').addClass('evenlistitem');
+            $('div.listitem:last-child').addClass('lastlistitem');
+            $('#locationtable tr:nth-child(odd)').addClass('oddtablerow');
+            $('.expandable').expander({
+               slicePoint:       600,              // make expandable if over this x chars
+               widow:            100,              // do not make expandable unless total length > slicePoint + widow
+               expandText:         '[read more]',  //text to use for expand link
+               expandEffect:     'fadeIn',         // or slideDown
+               expandSpeed:      700,              // in milliseconds
+               collapseTimer:    0,                // milliseconds before auto collapse; default is 0 (don't re-collape)
+               userCollapseText: '[collapse]'      // text for collaspe link
+            });
+         });
+
+         function js_highlighttoplink(selectedSpan)
+         {
+            $('.currentBrowseLink').toggleClass('browseLink').toggleClass('currentBrowseLink');
+            $(selectedSpan).toggleClass('currentBrowseLink');
+            $(selectedSpan).effect('highlight', {}, 300);
+         }
+
+         $(document).ready(function() {<?php echo($_ARCHON->PublicInterface->Header->OnLoad); ?>});
+         $(window).unload(function() {<?php echo($_ARCHON->PublicInterface->Header->OnUnload); ?>});
+         /* ]]> */
+      </script>
+      <?php
+      if($_ARCHON->PublicInterface->Header->Message && $_ARCHON->PublicInterface->Header->Message != $_ARCHON->Error)
+      {
+         $message = $_ARCHON->PublicInterface->Header->Message;
+      }
+      ?>
+   </head>
+   <body>
+      <?php
+      $_ARCHON->PublicInterface->outputGoogleAnalyticsCode();
+
+      if($message)
+      {
+         echo("<div class='message'>" . encode($message, ENCODE_HTML) . "</div>\n");
+      }
+      ?>
+      <div id="top">
+
+         <div id="logosearchwrapper">
+
+         
+            <div id="logo"><a href="http://archives.library.illinois.edu/sousa/" ><img src="<?php echo($_ARCHON->PublicInterface->ImagePath); ?>/logo.png" alt="logo" /></a> </div>
+
+            <div id="searchblock">
+               <form action="index.php" accept-charset="UTF-8" method="get" onsubmit="if(!this.q.value) { alert('Please enter search terms.'); return false; } else { return true; }">
+                  <div>
+                     <input type="hidden" name="p" value="core/search" />
+                     <span id="searchWord">Search:&nbsp;</span>
+	                     <select id="searchtype" name="setrepositoryid" required="">
+	                     	<option value="2">Sousa Archives Only</option>
+	                     	<option value="0">All University Archives</option>
+                     	</input>
+                     </select>
+                     <input type="text" size="31" title="search" maxlength="150" name="q" id="qfa" value="<?php echo(encode($_ARCHON->QueryString, ENCODE_HTML)); ?>" tabindex="100" />
+                     <input type="submit" value="Search" tabindex="300" class='button' title="Search" />
+                     <?php
+                     if(defined('PACKAGE_COLLECTIONS') && CONFIG_COLLECTIONS_SEARCH_BOX_LISTS)
+                     {
+                     ?>
+                        <input type="hidden" name="content" value="1" />
+                     <?php
+                     }
+                     ?>
+                  </div>
+               </form>
+            </div>
+         </div>
+
+         <div id="researchblock">
+            <?php
+                     if($_ARCHON->Security->isAuthenticated())
+                     {
+                       
+
+                        $logoutURI = preg_replace('/(&|\\?)f=([\\w])*/', '', $_SERVER['REQUEST_URI']);
+                        $Logout = (encoding_strpos($logoutURI, '?') !== false) ? '&amp;f=logout' : '?f=logout';
+                        $strLogout = encode($logoutURI, ENCODE_HTML) . $Logout;
+                        echo("<span class='bold'>Welcome, " . $_ARCHON->Security->Session->User->toString() . "<a href='$strLogout'>Logout</a>" . "</span><br/>");
+                   
+                     }
+                     elseif($_ARCHON->config->ForceHTTPS)
+                     {
+                        echo("<a href='index.php?p=core/login&amp;go='>Log In</a>");
+                     }
+                     else
+                     {
+                        echo("<a href='#' onclick='$(window).scrollTo(\"#archoninfo\"); if($(\"#userlogin\").is(\":visible\")) $(\"#loginlink\").html(\"Log In\"); else $(\"#loginlink\").html(\"Hide\"); $(\"#userlogin\").slideToggle(\"normal\"); $(\"#ArchonLoginField\").focus(); return false;'>Log In</a>");
+                     }
+
+                     if(!$_ARCHON->Security->userHasAdministrativeAccess())
+                     {
+                        $emailpage = defined('PACKAGE_COLLECTIONS') ? "collections/research" : "core/contact";
+
+                        echo(" | <a href='mailto:sousa@illinois.edu?subject=" . urlencode($_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI']) . "'>Contact Us</a>");
+                        if($_ARCHON->Security->isAuthenticated())
+                        {
+                           echo(" | <a href='?p=core/account&amp;f=account'>My Account</a>");
+                        }
+                        if(defined('PACKAGE_COLLECTIONS'))
+                        {
+                           $_ARCHON->Security->Session->ResearchCart->getCart();
+                           $EntryCount = $_ARCHON->Security->Session->ResearchCart->getCartCount();
+                           $class = $_ARCHON->Repository->ResearchFunctionality & RESEARCH_COLLECTIONS ? '' : 'hidewhenempty';
+                           $hidden = ($_ARCHON->Repository->ResearchFunctionality & RESEARCH_COLLECTIONS || $EntryCount) ? '' : "style='display:none'";
+
+                           echo("<span id='viewcartlink' class='$class' $hidden>| <a href='?p=collections/research&amp;f=cart&amp;referer=" . urlencode($_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI']) . "'>View Cart (<span id='cartcount'>$EntryCount</span>)</a></span>");
+                        }
+                     }
+            ?>
+                  </div>
+
+
+         <?php
+                     $arrP = explode('/', $_REQUEST['p']);
+                     $TitleClass = $arrP[0] == 'collections' && $arrP[1] != 'classifications' ? 'currentBrowseLink' : 'browseLink';
+                     $ClassificationsClass = $arrP[1] == 'classifications' ? 'currentBrowseLink' : 'browseLink';
+                     $SubjectsClass = $arrP[0] == 'subjects' ? 'currentBrowseLink' : 'browseLink';
+                     $CreatorsClass = $arrP[0] == 'creators' ? 'currentBrowseLink' : 'browseLink';
+                     $DigitalLibraryClass = $arrP[0] == 'digitallibrary' ? 'currentBrowseLink' : 'browseLink';
+         ?>
+                  <div id="browsebyblock">
+                     <span id="browsebyspan">
+                        Browse:
+                     </span>
+                     <span class="<?php echo($TitleClass); ?>">
+                        <a href="?p=collections/collections" onclick="js_highlighttoplink(this.parentNode); return true;">Collections</a>
+                     </span>
+                     <span class="<?php echo($DigitalLibraryClass); ?>">
+                        <a href="?p=digitallibrary/digitallibrary" onclick="js_highlighttoplink(this.parentNode); return true;">Digital Content</a>
+                     </span>
+                     <span class="<?php echo($SubjectsClass); ?>">
+                        <a href="?p=subjects/subjects" onclick="js_highlighttoplink(this.parentNode); return true;">Subjects</a>
+                     </span>
+                     <span class="<?php echo($CreatorsClass); ?>">
+                        <a href="?p=creators/creators" onclick="js_highlighttoplink(this.parentNode); return true;">Creators</a>
+                     </span>
+                     <span class="<?php echo($ClassificationsClass); ?>">
+                        <a href="?p=collections/classifications" onclick="js_highlighttoplink(this.parentNode); return true;">Record Groups</a>
+                     </span>
+                  </div>
+
+
+                  </div>
+                  
+                    </div>
+            <div id="breadcrumbblock"><span class='bold'>Location: </span><?php echo($_ARCHON->PublicInterface->createNavigation()); ?></div>
+            <div id="breadcrumbclearblock">.</div>
+
+      <script type="text/javascript">
+         /* <![CDATA[ */
+         if ($.browser.msie && parseInt($.browser.version, 10) <= 8){
+            $.getScript('packages/core/js/jquery.corner.js', function(){                                                              
+               $("#searchblock").corner("5px");
+               $("#browsebyblock").corner("tl 10px");
+                         
+               $(function(){
+                  $(".bground").corner("0px");
+                  $(".mdround").corner("0px");
+                  $(".smround").corner("0px");
+                  $("#dlsearchblock").corner("bottom 10px");
+               });
+            });
+         }
+         /* ]]> */
+      </script>
+      <div id="main">
